Pietrzak no-std
Removed all parts of fastcrypto not needed by pietrzak VDF and made
changes to allow it to be used from no-std environments.

Mostly it's removing things, or changing dependency specifications, but
the hash for `r` calculations had to change as well. This means that
verification can't be guaranteed to work with proofs generated with the
original fastcrypto library.